resizeimage_rest/serializers.py
from rest_framework import serializers
from resizeimage.models import Image

class ImageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Image
        fields = ('pk', 'image',)
