requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
kombu==4.6.3
Markdown==3.1.1
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
vine==1.3.0
